linkedin_analyser--v1/linkedin/requirements.txt
fastapi
requests
python-dotenv
fastapi
uvicorn
requests
pydantic
langchain_google_genai